RavenDB-20732 Avoid leaking change vector from shard to shard
For conflicts, attachment, revisions and deletion we had a code path that leaked change vectors
We must avoid this, so features like subscriptions/replication/ETL will not skip documents after resharding and failover

+++ test/Tests.Infrastructure/RavenTestBase.cs
@@ -287,12 +287,19 @@ namespace FastTests
                             if (CreatedStores.TryRemove(store) == false)
                                 return; // can happen if we are wrapping the store inside sharded one
 
+                            if (sharded && options.CreateDatabase)
+                            {
+                                if (Servers.Contains(serverToUse) || IsGlobalOrLocalServer(serverToUse))
+                                {
+                                    AsyncHelpers.RunSync(() => Sharding.EnsureNoDatabaseChangeVectorLeakAsync(store.Database));
+                                }
+                            }
+
                             DeleteDatabaseResult result = null;
-                            if (options.DeleteDatabaseOnDispose)
+                            if (options.DeleteDatabaseOnDispose && options.CreateDatabase)
                             {
                                 result = DeleteDatabase(options, serverToUse, name, hardDelete, store);
                             }
-
                             if (Servers.Contains(serverToUse) && result != null)
                             {
                                 var timeout = options.DeleteTimeout ?? TimeSpan.FromSeconds(Debugger.IsAttached ? 150 : 15);
